+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Up Next
 -------------
 - Fix a bug where `quilt setup-tls` would fail when writing to a directory whose
 parent does not exist.
+- Auto-generate TLS credentials when starting the daemon if the provided
+credentials don't exist.
 
 JavaScript API-breaking changes:
 - Remove the Container.replicate() method. Users should create multiple